# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Payment Relay API
  version: 0.3.0
  description: |
    Internal payment relay — systems, wallets, payouts, invoices, reports, and pay page.

    **Auth**: Protected routes require `X-API-Key: sk_live_...` header.

    **Amounts**: All monetary values are in minor units (cents).
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /auth/login:
    post:
      tags: [systems]
      summary: Merchant dashboard login (username/password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Session token for X-Session-Token
        '401':
          description: Invalid credentials
...
  /systems:
    post:
      tags: [systems]
      summary: Register a system (auto-seeds wallets)
      operationId: createSystem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSystemRequest'
      responses:
        '200':
          description: System created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSystemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Prefix already exists
  /systems/{id}:
    get:
      tags: [systems]
      summary: Get system by ID
      operationId: getSystem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: System details (no API key in response)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemPublic'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments:
    post:
      tags: [payments]
      summary: Process a payout
      operationId: processPayment
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
      responses:
        '200':
          description: Payment result (including idempotent replay)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessPaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          description: Insufficient wallet balance
        '403':
          description: Country not enabled for system
        '409':
          description: Idempotency key reused with different request body
  /payments/{id}:
    get:
      tags: [payments]
      summary: Get payment by relay ID
      operationId: getPayment
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessPaymentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /batches:
    post:
      tags: [batches]
      summary: Create and process a payout batch (sync, max 100 lines)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [system_id, idempotency_key, lines]
              properties:
                system_id: { type: string, format: uuid }
                idempotency_key: { type: string }
                lines:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items:
                    type: object
                    required: [amount, currency, country, payment_method]
                    properties:
                      amount: { type: integer }
                      currency: { type: string }
                      country: { type: string }
                      external_id: { type: string }
                      payment_method:
                        $ref: '#/components/schemas/PaymentMethod'
      responses:
        '200':
          description: Batch result (may be partial)
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Idempotency conflict
  /batches/{id}:
    get:
      tags: [batches]
      summary: Get payout batch by ID
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Batch with lines
        '404':
          $ref: '#/components/responses/NotFound'
  /wallets/{system_id}:
    get:
      tags: [wallets]
      summary: List wallets for a system
      operationId: listWallets
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/SystemId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wallet'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: API key does not match system_id
  /transactions/{system_id}:
    get:
      tags: [payments]
      summary: List transactions for a system
      operationId: listTransactions
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/SystemId'
        - name: external_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
  /invoices:
    post:
      tags: [invoices]
      summary: Create invoice with QR code
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '200':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
    get:
      tags: [invoices]
      summary: List invoices
      security:
        - ApiKeyAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, paid, expired, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceResponse'
  /invoices/reference/{reference}:
    get:
      tags: [invoices]
      summary: Get invoice by reference
      security:
        - ApiKeyAuth: []
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
  /invoices/{id}/collect:
    post:
      tags: [invoices]
      summary: Collect payment for open invoice
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectInvoiceRequest'
      responses:
        '200':
          description: Invoice paid, wallet credited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessPaymentResponse'
  /invoices/{id}/cancel:
    post:
      tags: [invoices]
      summary: Cancel open invoice
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Invoice cancelled
  /invoices/{id}/refund:
    post:
      tags: [invoices]
      summary: Refund a paid invoice (full or partial)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, idempotency_key]
              properties:
                amount: { type: integer }
                idempotency_key: { type: string }
                phone: { type: string }
                provider: { type: string }
      responses:
        '200':
          description: Refund result
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: Insufficient wallet balance
  /reports/transactions:
    get:
      tags: [reports]
      summary: Transaction report
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
        - $ref: '#/components/parameters/ReportFormat'
        - name: status
          in: query
          schema:
            type: string
        - name: detail
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: JSON summary or CSV export
  /reports/wallets:
    get:
      tags: [reports]
      summary: Wallet balance report
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
        - $ref: '#/components/parameters/ReportFormat'
      responses:
        '200':
          description: JSON summary or CSV export
  /reports/invoices:
    get:
      tags: [reports]
      summary: Invoice report
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
        - $ref: '#/components/parameters/ReportFormat'
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: JSON summary or CSV export
  /webhook-endpoints:
    get:
      tags: [webhooks]
      summary: List webhook endpoints for the authenticated system
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookEndpoint'
    post:
      tags: [webhooks]
      summary: Add a webhook endpoint
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookEndpointRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: URL already registered
  /webhook-endpoints/{id}:
    patch:
      tags: [webhooks]
      summary: Update a webhook endpoint
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookEndpointRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpoint'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [webhooks]
      summary: Delete a webhook endpoint
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/login:
    post:
      tags: [admin]
      summary: Sign in with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Session token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  username:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
        '401':
          description: Invalid credentials
  /admin/logout:
    post:
      tags: [admin]
      summary: End the current admin session
      security:
        - AdminTokenAuth: []
      responses:
        '200':
          description: Logged out
  /admin/me:
    get:
      tags: [admin]
      summary: Current admin profile
      security:
        - AdminTokenAuth: []
      responses:
        '200':
          description: Admin user
        '401':
          description: Missing or invalid session
  /admin/systems:
    get:
      tags: [admin]
      summary: List all tenant systems
      security:
        - AdminTokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminSystemSummary'
        '401':
          description: Invalid or missing admin session
  /admin/systems/{id}:
    get:
      tags: [admin]
      summary: Get system detail with wallets and webhooks
      security:
        - AdminTokenAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSystemDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /pay/{reference}:
    get:
      tags: [pay-page]
      summary: Serve React pay page shell
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: React SPA HTML shell
          content:
            text/html:
              schema:
                type: string
        '503':
          description: Frontend not built
  /api/pay/{reference}:
    get:
      tags: [pay-page]
      summary: Load invoice for pay page
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayPageResponse'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [pay-page]
      summary: Submit payment from pay page
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaySubmitRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaySubmitResponse'
        '429':
          description: Rate limit exceeded
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "System API key from POST /systems (format: sk_live_...)"
    AdminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Session token from POST /admin/login (Bearer also accepted)
  parameters:
    SystemId:
      name: system_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReportFrom:
      name: from
      in: query
      required: true
      schema:
        type: string
        format: date-time
    ReportTo:
      name: to
      in: query
      required: true
      schema:
        type: string
        format: date-time
    ReportFormat:
      name: format
      in: query
      schema:
        type: string
        enum: [json, csv]
        default: json
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
    WalletSeedOverride:
      type: object
      required: [country, currency, amount]
      properties:
        country:
          type: string
        currency:
          type: string
        amount:
          type: integer
          format: int64
    CreateSystemRequest:
      type: object
      required: [name, prefix, enabled_countries, username, password]
      properties:
        name:
          type: string
        prefix:
          type: string
          pattern: '^[A-Z0-9]{2,8}$'
        username:
          type: string
          description: Dashboard login username
        password:
          type: string
          format: password
          description: Dashboard login password (min 8 chars)
        enabled_countries:
          type: array
          items:
            type: string
        webhook_url:
          type: string
          format: uri
        wallet_seeds:
          type: array
          items:
            $ref: '#/components/schemas/WalletSeedOverride'
    CreateSystemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prefix:
          type: string
        username:
          type: string
        api_key:
          type: string
        session_token:
          type: string
        wallets_seeded:
          type: integer
    SystemPublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prefix:
          type: string
        enabled_countries:
          type: array
          items:
            type: string
        webhook_url:
          type: string
          nullable: true
    PaymentMethod:
      type: object
      required: [type, details]
      properties:
        type:
          type: string
          enum: [mmo, card, bank]
        details:
          type: object
          additionalProperties: true
    ProcessPaymentRequest:
      type: object
      required: [system_id, external_id, amount, currency, country, payment_method, idempotency_key]
      properties:
        system_id:
          type: string
          format: uuid
        external_id:
          type: string
        amount:
          type: integer
          format: int64
        currency:
          type: string
        country:
          type: string
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
        idempotency_key:
          type: string
    ProcessPaymentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        external_id:
          type: string
        status:
          type: string
        gateway_reference:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        system_id:
          type: string
          format: uuid
        country:
          type: string
        currency:
          type: string
        balance:
          type: integer
          format: int64
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        system_id:
          type: string
          format: uuid
        external_id:
          type: string
        amount:
          type: integer
        currency:
          type: string
        country:
          type: string
        status:
          type: string
        gateway:
          type: string
        created_at:
          type: string
          format: date-time
    CreateInvoiceRequest:
      type: object
      required: [amount, currency, country]
      properties:
        amount:
          type: integer
          format: int64
        currency:
          type: string
        country:
          type: string
        description:
          type: string
        expires_in_hours:
          type: integer
          default: 24
    InvoiceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
        system_id:
          type: string
          format: uuid
        amount:
          type: integer
        currency:
          type: string
        country:
          type: string
        status:
          type: string
          enum: [open, paid, expired, cancelled]
        description:
          type: string
        expires_at:
          type: string
          format: date-time
        paid_at:
          type: string
          format: date-time
          nullable: true
        transaction_id:
          type: string
          format: uuid
          nullable: true
        qr_url:
          type: string
        qr_code_png_base64:
          type: string
    CollectInvoiceRequest:
      type: object
      required: [payment_method, idempotency_key]
      properties:
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
        idempotency_key:
          type: string
    PayPageResponse:
      type: object
      properties:
        reference:
          type: string
        amount:
          type: integer
        amount_display:
          type: string
        currency:
          type: string
        country:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        expires_at:
          type: string
          format: date-time
        payable:
          type: boolean
        form_token:
          type: string
        idempotency_key:
          type: string
        providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderOption'
    ProviderOption:
      type: object
      properties:
        value:
          type: string
        label:
          type: string
    PaySubmitRequest:
      type: object
      required: [phone, provider, idempotency_key, form_token]
      properties:
        phone:
          type: string
        provider:
          type: string
        idempotency_key:
          type: string
        form_token:
          type: string
    PaySubmitResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        amount_display:
          type: string
        reference:
          type: string
    WebhookEndpoint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        system_id:
          type: string
          format: uuid
        url:
          type: string
        label:
          type: string
          nullable: true
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateWebhookEndpointRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
        label:
          type: string
    UpdateWebhookEndpointRequest:
      type: object
      properties:
        url:
          type: string
        label:
          type: string
        enabled:
          type: boolean
    AdminSystemSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prefix:
          type: string
        enabled_countries:
          type: array
          items:
            type: string
        webhook_url:
          type: string
          nullable: true
        webhook_endpoints:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    AdminSystemDetail:
      type: object
      properties:
        system:
          $ref: '#/components/schemas/SystemPublic'
        wallets:
          type: array
          items:
            $ref: '#/components/schemas/Wallet'
        webhook_endpoints:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEndpoint'